agent/slack_utils.py
import os
import datetime
from slack_sdk import WebClient
from dotenv import load_dotenv

load_dotenv()

client = WebClient(token=os.getenv("SLACK_TOKEN"))


with open("rca.txt") as f:
    analysis = f.read()

message = f"""
🚨 *High CPU Alert*
Root Cause: {analysis}
✅ Action: Auto-restarted container
🕒 Timestamp: {datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S')}
"""

client.chat_postMessage(channel="#alerts", text=message)
