requirements.txt
flask
openai
slack_sdk
prometheus-api-client
python-dotenv
streamlit
